Added logic to double-up any embedded text delimiter characters
for example, the text:

this is "sample text

Would error while being parsed by other programs when it sees the string, and enclosed by double quotes:
"this is "sample text"

now, the library will correctly output this:

"this is ""sample text"

Which should parse back out to the original string with the double quote intact.
FileHelpers will correctly parse this.